requirements.txt
visca-over-ip==0.5.1
gpiozero==2.0.1
PyYAML==6.0.2
PyQt5==5.15.11
pygame==2.6.1